Refactor: reformat code for consistency and readability across components
- Expand single-line constructors to multi-line for UserControl views
- Reorder and align field declarations in view models and state classes
- Use consistent brace and indentation style in methods and property accessors
- Move helper methods and address computation for clarity in KatanaAddressMap
- Adjust parameter ordering and whitespace in method calls and initializers
- Update comments and summaries for clarity
- Minor logic and formatting tweaks in utility and service classes

+++ tests/Kataka.Tests/RolandSysExBuilderTests.cs
@@ -151,9 +151,9 @@ public sealed class RolandSysExBuilderTests
             KatanaMkIIParameterCatalog.ModType.SkippedValues);
         Assert.Equal(KatanaMkIIParameterCatalog.ModType.SkippedValues, KatanaMkIIParameterCatalog.FxType.SkippedValues);
 
-        Assert.Equal((byte)10, KatanaMkIIParameterCatalog.DelayType.Maximum);
-        Assert.Equal((byte)10, KatanaMkIIParameterCatalog.Delay2Type.Maximum);
-        Assert.Equal((byte)6, KatanaMkIIParameterCatalog.ReverbType.Maximum);
+        Assert.Equal(10, KatanaMkIIParameterCatalog.DelayType.Maximum);
+        Assert.Equal(10, KatanaMkIIParameterCatalog.Delay2Type.Maximum);
+        Assert.Equal(6, KatanaMkIIParameterCatalog.ReverbType.Maximum);
     }
 
     [Fact]
